// MochiR.Client/src/app/api/models/self-profile-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export type SelfProfileDto = {
'id': string;
'userName': string | null;
'displayName': string | null;
'email': string | null;
'emailConfirmed': boolean;
'phoneNumber': string | null;
'phoneNumberConfirmed': boolean;
'avatarUrl': string | null;
'twoFactorEnabled': boolean;
'lockoutEnabled': boolean;
'lockoutEnd': string | null;
'createdAtUtc': string;
'followersCount': number;
'followingCount': number;
'isAdmin': boolean;
};